requirements.txt
amqp==2.2.2
arrow==0.10.0
billiard==3.5.0.3
bs4==0.0.1
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
DateTime==4.2
Flask==0.12.2
Flask-Compress==1.4.0
gspread==0.6.2
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
numpy==1.13.2
oauth2client==4.1.2
oauthlib==2.0.4
pandas==0.20.3
psycopg2==2.7.3.1
pyasn1==0.3.6
pyasn1-modules==0.1.4
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
setupfiles==0.0.50
six==1.11.0
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
zope.interface==4.4.3
google-cloud-bigquery==0.28.0
google-cloud-logging==1.4.0
bingads==11.5.6
